Split up `Section` for non control point use
