+++ osu.Game/Screens/Edit/Section.cs
@@ -1,39 +1,29 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Overlays;
 using osuTK;
 
-namespace osu.Game.Screens.Edit.Timing
+namespace osu.Game.Screens.Edit
 {
-    internal abstract class Section<T> : CompositeDrawable
-        where T : ControlPoint
+    internal abstract class Section : CompositeDrawable
     {
-        private OsuCheckbox checkbox;
-        private Container content;
+        protected Container Content;
+        protected OsuCheckbox Checkbox;
 
         protected FillFlowContainer Flow { get; private set; }
 
-        protected Bindable<T> ControlPoint { get; } = new Bindable<T>();
-
         private const float header_height = 50;
 
-        [Resolved]
-        protected EditorBeatmap Beatmap { get; private set; }
-
-        [Resolved]
-        protected Bindable<ControlPointGroup> SelectedGroup { get; private set; }
-
-        [Resolved(canBeNull: true)]
-        protected IEditorChangeHandler ChangeHandler { get; private set; }
+        /// <summary>
+        /// The name of this section, as seen from the user interface.
+        /// </summary>
+        protected abstract string SectionName { get; }
 
         [BackgroundDependencyLoader]
         private void load(OverlayColourProvider colours)
@@ -54,15 +44,15 @@ namespace osu.Game.Screens.Edit.Timing
                     Padding = new MarginPadding { Horizontal = 10 },
                     Children = new Drawable[]
                     {
-                        checkbox = new OsuCheckbox
+                        Checkbox = new OsuCheckbox
                         {
                             Anchor = Anchor.CentreLeft,
                             Origin = Anchor.CentreLeft,
-                            LabelText = typeof(T).Name.Replace(nameof(Beatmaps.ControlPoints.ControlPoint), string.Empty)
+                            LabelText = SectionName
                         }
                     }
                 },
-                content = new Container
+                Content = new Container
                 {
                     Y = header_height,
                     RelativeSizeAxes = Axes.X,
@@ -91,42 +81,10 @@ namespace osu.Game.Screens.Edit.Timing
         {
             base.LoadComplete();
 
-            checkbox.Current.BindValueChanged(selected =>
+            Checkbox.Current.BindValueChanged(selected =>
             {
-                if (selected.NewValue)
-                {
-                    if (SelectedGroup.Value == null)
-                    {
-                        checkbox.Current.Value = false;
-                        return;
-                    }
-
-                    if (ControlPoint.Value == null)
-                        SelectedGroup.Value.Add(ControlPoint.Value = CreatePoint());
-                }
-                else
-                {
-                    if (ControlPoint.Value != null)
-                    {
-                        SelectedGroup.Value.Remove(ControlPoint.Value);
-                        ControlPoint.Value = null;
-                    }
-                }
-
-                content.BypassAutoSizeAxes = selected.NewValue ? Axes.None : Axes.Y;
+                Content.BypassAutoSizeAxes = selected.NewValue ? Axes.None : Axes.Y;
             }, true);
-
-            SelectedGroup.BindValueChanged(points =>
-            {
-                ControlPoint.Value = points.NewValue?.ControlPoints.OfType<T>().FirstOrDefault();
-                checkbox.Current.Value = ControlPoint.Value != null;
-            }, true);
-
-            ControlPoint.BindValueChanged(OnControlPointChanged, true);
         }
-
-        protected abstract void OnControlPointChanged(ValueChangedEvent<T> point);
-
-        protected abstract T CreatePoint();
     }
 }
